Add files via upload
Extracting Amazon Stock data
A company's stock share is a piece of the company more precisely:
A stock (also known as equity) is a security that represents the ownership of a fraction of a corporation. This entitles the owner of the stock to a proportion of the corporation's assets and profits equal to how much stock they own. Units of stock are called "shares."
An investor can buy a stock and sell it later. If the stock price increases, the investor profits, If it decreases,the investor with incur a loss. Determining the stock price is complex; it depends on the number of outstanding shares, the size of the company's future profits, and much more. People trade stocks throughout the day the stock ticker is a report of the price of a certain stock, updated continuously throughout the trading session by the various stock market exchanges.
You are a data scientist working for a hedge fund; it's your job to determine any suspicious stock activity. In this file you will extract stock data using a Python library. We will use the requests library, it allows us to extract data for stocks returning data in a pandas dataframe.Here,I am choosing Amazon stock data for webscraping.